+++ keyboards/sofle/keymaps/choc-brian/sofle.json
@@ -1,0 +1,255 @@
+{
+  "name": "Sofle",
+  "vendorProductId": 4231135879,
+  "macros": ["", "", "", "", "", "", "", "", "", "", "", "", "", "", "", ""],
+  "layers": [
+    [
+      "KC_GRV",
+      "KC_1",
+      "KC_2",
+      "KC_3",
+      "KC_4",
+      "KC_5",
+      "KC_ESC",
+      "KC_Q",
+      "KC_W",
+      "KC_E",
+      "KC_R",
+      "KC_T",
+      "KC_TAB",
+      "KC_A",
+      "KC_S",
+      "KC_D",
+      "KC_F",
+      "KC_G",
+      "KC_LSFT",
+      "KC_Z",
+      "KC_X",
+      "KC_C",
+      "KC_V",
+      "KC_B",
+      "KC_LGUI",
+      "KC_LALT",
+      "KC_LCTL",
+      "0x5d5e",
+      "KC_ENT",
+      "KC_MUTE",
+      "KC_MINS",
+      "KC_0",
+      "KC_9",
+      "KC_8",
+      "KC_7",
+      "KC_6",
+      "KC_BSPC",
+      "KC_P",
+      "KC_O",
+      "KC_I",
+      "KC_U",
+      "KC_Y",
+      "KC_QUOT",
+      "KC_SCLN",
+      "KC_L",
+      "KC_K",
+      "KC_J",
+      "KC_H",
+      "KC_RSFT",
+      "KC_SLSH",
+      "KC_DOT",
+      "KC_COMM",
+      "KC_M",
+      "KC_N",
+      "KC_RGUI",
+      "KC_RALT",
+      "KC_RCTL",
+      "0x5d5f",
+      "KC_SPC",
+      "KC_NO"
+    ],
+    [
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_INS",
+      "KC_PSCR",
+      "KC_APP",
+      "KC_NO",
+      "KC_NO",
+      "KC_TRNS",
+      "KC_LALT",
+      "KC_LCTL",
+      "KC_LSFT",
+      "KC_NO",
+      "KC_CAPS",
+      "KC_TRNS",
+      "KC_UNDO",
+      "KC_CUT",
+      "KC_COPY",
+      "KC_PASTE",
+      "KC_NO",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_BSPC",
+      "0x5d65",
+      "0x5d62",
+      "KC_UP",
+      "0x5d61",
+      "KC_PGUP",
+      "KC_BSPC",
+      "KC_DEL",
+      "KC_RGHT",
+      "KC_DOWN",
+      "KC_LEFT",
+      "KC_PGDN",
+      "KC_TRNS",
+      "KC_NO",
+      "0x5d64",
+      "KC_NO",
+      "0x5d63",
+      "KC_NO",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS"
+    ],
+    [
+      "KC_TRNS",
+      "KC_F1",
+      "KC_F2",
+      "KC_F3",
+      "KC_F4",
+      "KC_F5",
+      "KC_GRV",
+      "KC_1",
+      "KC_2",
+      "KC_3",
+      "KC_4",
+      "KC_5",
+      "KC_TRNS",
+      "KC_EXLM",
+      "KC_AT",
+      "KC_HASH",
+      "KC_DLR",
+      "KC_PERC",
+      "KC_TRNS",
+      "KC_EQL",
+      "KC_MINS",
+      "KC_PLUS",
+      "KC_LCBR",
+      "KC_RCBR",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_F11",
+      "KC_F10",
+      "KC_F9",
+      "KC_F8",
+      "KC_F7",
+      "KC_F6",
+      "KC_F12",
+      "KC_0",
+      "KC_9",
+      "KC_8",
+      "KC_7",
+      "KC_6",
+      "KC_PIPE",
+      "KC_RPRN",
+      "KC_LPRN",
+      "KC_ASTR",
+      "KC_AMPR",
+      "KC_CIRC",
+      "KC_TRNS",
+      "KC_BSLS",
+      "KC_COLN",
+      "KC_SCLN",
+      "KC_RBRC",
+      "KC_LBRC",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS"
+    ],
+    [
+      "KC_NO",
+      "RGB_M_B",
+      "RGB_M_R",
+      "RGB_M_SW",
+      "RGB_M_SN",
+      "RGB_M_K",
+      "RESET",
+      "RGB_MOD",
+      "RGB_HUI",
+      "RGB_SAI",
+      "RGB_VAI",
+      "RGB_SPI",
+      "KC_NO",
+      "RGB_RMOD",
+      "RGB_HUD",
+      "RGB_SAD",
+      "RGB_VAD",
+      "RGB_SPD",
+      "0x5d00",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "RGB_M_G",
+      "RGB_M_X",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_VOLU",
+      "KC_MUTE",
+      "KC_VOLD",
+      "KC_NO",
+      "KC_NO",
+      "KC_NO",
+      "KC_MNXT",
+      "KC_MPLY",
+      "KC_MPRV",
+      "KC_NO",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_TRNS",
+      "KC_NO"
+    ]
+  ]
+}
